Base da integração no github criada
ainda não está integrada com dados do languim, ainda não executa quando usa a extensao do made

ja cria issue no repositorio com assignees, labels, title e body, e coloca esse issue num novo project

o token, nome do repositorio e o nome da organização esta sendo declarada no .env

para testar tem que instalar o axios e o dotenv, e para executar roda um npm run start:github

(todos os dados são estaticos atualmente, precisa colocar os do .made)
