+++ .env
@@ -1,3 +1,3 @@
-GITHUB_TOKEN=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
+GITHUB_TOKEN=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
 ORGANIZATION=PS-MADE
 REPOSITORY_NAME=testeProject
