+++ package.json
@@ -15,7 +15,7 @@
     },
     "type": "module",
     "scripts": {
-        "start": "node --no-warnings --loader ts-node/esm src/cli/project_management/application/github_issues/application/main.ts",
+        "start:github": "node --no-warnings --loader ts-node/esm src/cli/project_management/application/github_issues/application/main.ts",
         "build": "tsc -b tsconfig.src.json && node esbuild.mjs",
         "watch": "concurrently -n tsc,esbuild -c blue,yellow \"tsc -b tsconfig.src.json --watch\" \"node esbuild.mjs --watch\"",
         "lint": "eslint src --ext ts",
@@ -49,6 +49,7 @@
         "monaco-editor-workers": "~0.44.0",
         "monaco-editor-wrapper": "~3.5.0",
         "monaco-languageclient": "~7.2.0",
+        "octokit": "^4.1.3",
         "ts-node": "^10.9.2",
         "vscode-languageclient": "~9.0.1",
         "vscode-languageserver": "~9.0.1"
