+++ src/cli/project_management/application/github_issues/application/main.ts
@@ -1,6 +1,6 @@
-import { createProject } from './createProject.js';
+import { createProject, addIssueToProject } from './createProject.js';
 import { createIssue } from './createIssue.js';
-import { addIssueToProject } from './addIssueToProject.js';
+
 
 const ORGANIZATION = 'PS-MADE';
 const PROJECT_TITLE = 'PROJETOTESTANDOPROMADE';
@@ -8,25 +8,40 @@ const REPOSITORY_NAME = 'testeProject';
 
 async function main() {
     try {
+        
         console.log('Iniciando criação do projeto...');
         const projectId = await createProject(ORGANIZATION, PROJECT_TITLE);
         console.log('Projeto criado com ID:', projectId);
 
         console.log('Iniciando criação da issue...');
-        const issueId = await createIssue(
+
+        /*
+            organizationName: string,
+            repositoryName: string,
+            title: string,
+            body: string,
+            labels: string[],
+            assignees: string[],
+        */
+       
+        const issue = await createIssue(
             ORGANIZATION,
             REPOSITORY_NAME,
-            'Minha primeira issue',
-            'Descrição da minha primeira issue',
+            'Nome Issue',
+            'Descrição da issue',
             ['bug', 'duplicate'],
-            ['jonathancastrosilva'],
-            'BUG'
+            ['jonathancastrosilva', 'JosiasNJB'],
         );
         
-        console.log('Issue criada com ID:', issueId);
+        console.log('Issue criada com ID:', issue.id);
+        console.log('✅ Todos os assignees foram adicionados com sucesso.');
 
         console.log('Adicionando issue ao projeto...');
-        await addIssueToProject(projectId, issueId.id);
+        await addIssueToProject(projectId, issue.id);
+
+        //await addFieldToIssue(projectId, itemId, 'Type', 'Bug');
+
+
         console.log('✅ Issue adicionada ao projeto.');
     } catch (error: any) {
         console.error('❌ Erro:', error.response?.data || error.message);
