+++ src/cli/project_management/application/github_issues/utils/axiosInstance.ts
@@ -8,8 +8,9 @@ const GITHUB_TOKEN = process.env.GITHUB_TOKEN || '';
 const GITHUB_API_URL = 'https://api.github.com/graphql';
 
 if (!GITHUB_TOKEN) {
-    throw new Error('GITHUB_TOKEN não está definido. Configure-o como uma variável de ambiente.');
+    throw new Error('❌ GITHUB_TOKEN não está definido. Configure-o como uma variável de ambiente.');
 }
+
 
 export const axiosInstance = axios.create({
     baseURL: GITHUB_API_URL,
